pyproject.toml
[project]
name = "agent-dev-adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.15.1",
    "google-generativeai>=0.8.5",
    "litellm>=1.77.5",
    "psutil>=7.1.0",
    "python-dotenv>=1.1.1",
    "yfinance>=0.2.66",
]
